bcache: add id and reserve fields to btree_op
Previously functions that operated on btrees would take both a btree_op and
a btree_id. Adding the id to btree_op simplifies parameter lists. It also
allows the caller to pass in the btree node reserve to use for allocations.

This is cleaner than having a moving_gc bit inside btree_op, and with a
future patch that adds a tiering reserve it eliminates the need to have a
separate tiering bit.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>