bcache: add new sectors before subtracting old sectors
Otherwise, when we insert a new copy of an existing key,
we unmark the old bucket before we remark the old bucket,
or mark the new bucket. If the old bucket is invalidated
in this time window, we will hit the stale dirty pointer
assertion and lose data.

The only tricky case is when replace_key != NULL, where
the size of the inserted key depends on which existing
keys it overlaps with, which have already been subtracted.

To deal with this, we always add the insert key at the
start, and only subtract the parts that didn't match the
replace key at the end.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>