bcache: remove dead code in allocator thread
We should never get to a point where we're completely out of buckets,
because we now block foreground writes once free buckets crosses a
certain threshold. As a result, the code for cranking up the tiering
and moving GC rate would never be executed, so just nuke it.

We do have to wake up btree GC when the number of buckets whose gen
we cannot increment crosses a threshold, though, do that wake up
when we are looking for buckets to invalidate in alloc.c.

The saturated sector count case would already wake up the allocator,
which would wake up btree GC if the allocator was waiting for buckets.
Now, wake up btree GC directly. This is a change in behavior and a
potential performance regression, since now we run btree GC whenever
we subtract free_inc_reserve worth of saturated sectors, and not when
we are also out of space.  However, we shouldn't be hitting this code
path in our testing yet, and it needs tuning anyway.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>