bcache: Avoid memmove when dequeuing keylist entries.  Use pointers.
bch_keylist_pop_front used to memmove from back to front to dequeue
an element.  This leads to quadratic time if a keylist is long.

Add another pointer to the current front, and bump that instead.

Rename the fields of a keylist with changed semantics to catch all users.

Add a pointer to the end of the currently-allocated array.

Add a few asserts (BUG_ON).

Signed-off-by: Bill Rozas <gjr@datera.io>
Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>