bcache: write out prios even if CACHE_SYNC is false
We don't use CACHE_SYNC for its original purpose anymore (where we don't
preserve cached data across reboots). Instead, it is used to indicate a
fresh cache set that has just come out of make-bcache.

In this case, we definitely still want to write out prios and gens when
CACHE_SYNC is false, since it only gets set to true after we've started
the allocator threads and after we've allocated btree node roots.

The symptom would be that by not writing the gens for the first free_inc
batch, we would hit stale pointers on reboot when the change to make the
stale pointer check stricter was applied (using != instead of gen_after()).

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>