bcache: remove bogus BUG_ON()
If we failed to reallocate memory for the keylist, we would bail out
of the refill path, issue some IO, and continue refilling once we had
space in the keylist again. In this case, the total count of sectors
added to the keylist in this round could exceed the number of sectors
currently in the keylist, triggering the BUG_ON().

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>