bcache: fix handling of PTR_LOST_DEV keys
If we lose all copies of a key, we want to fail the read, not
treat it as a hole in the keyspace, so don't mark the key as
deleted. Instead, set the key type to error.

This fixes a regression from "New bkey format".

Also, instead of having bch_extent_normalize() set a key's
deleted flag, just return true if the key should be dropped.
If the key is an extent and has no pointers, it becomes a
discard.

This could come up in bch_flag_key_bad(). If a cached key
points to a device that has gone bad, we end up dropping all
pointers from the key. This would cause us to insert a deleted
key, which triggers a BUG_ON ever since "Don't insert deleted
keys with nonzero size". What we want instead is to discard
this range of the keyspace -- we just lost some cached data,
which is not an error.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>