bcache: enforce minimum journal entry size
This lowers write latency by reducing the likelyhood that we fill
up a journal entry really quickly while the previous journal write
is still in progress. Previously, this would often happen when we
are near the end of a journal bucket. Now, we just skip to the
next journal bucket if we can't get a 32KiB journal entry.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>