bcache: Cleanup debug parameters
The main point of this patch was to add module parameters for all the debug
parameters - so that if something like e.g. xfstests is creating/registering a
cache set, we have a global parameter we can flip on instead of hacking
xfstests.

Also fix bch_count_data(), which was using int instead of s64 (since the size
field of struct bkey is now 32 bits, that easily overflows if discards are
involved).