bcache: Journal when we insert keys, not before
Previously, journalling an index update and doing the index update were two
different operations. This was problematic even in the current code, and was
going to be a major issue for future work; basically, any index update that
required traversing and locking the btree before we know what's actually being
done (which currently includes replace operations) couldn't make use of the
journal.

Now, any index update that uses bch_btree_insert_node() gets journalled - i.e.
everything (at least to leaf nodes, for now).

This also means the order that index updates happen in is preserved in the
journal, which was a bug waiting to happen (if it wasn't a bug already).

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>