bcache: better iterator invalidation
Previously, if a btree map fn invalidated an iterator, it would force
bch_btree_map_keys()/bch_btree_map_nodes() to redo the lookup by returning
-EINTR.

But we're going to rework bch_btree_map_keys() to make sure it passes every key
to the map fn precisely once, so we need to differentiate between cases where
the map fn does need to be called again (for which it will still return -EINTR)
from cases where it's done with that key, but the iterator was still
invalidated.

This also means the map_fn doesn't need to know about bch_btree_insert_node() or
soemthing else it calls invalidating the iterator.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>