bcache: MAP_END_KEY -> MAP_HOLES
MAP_END_KEY was created for cases (like cache_lookup_fn) where if we _don't_
find the key we're looking for, we still need to do something with that btree
node.

This has since come up in various places (inode/dirent creation in particular),
but it was only just now that I realized that in all these cases what we really
want to do is iterate over the _keyspace_, not just the keys that happen to be
present.

So, that's what MAP_HOLES does now - the code that implements it is gross, but
it _considerably_ simplifies all the users.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>