bcache: Use fewer workqueues
With the recent stuff to make allocation all asynchronous, we now shouldn't need
different workqueues for tiering, copy gc, etc. - so consolidate down to a
single workqueue.

This ought to help performance a bit, as previously a write request would run
out of two different workqueues for different parts of the write path, now we'll
only be using a single worker thread.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>