bcache: Safer check for if locks need to be upgraded before split
Using btree_node_root() to get the current depth of the btree when the root
might not be locked wasn't entirely safe.

But b->parent should always be non null for a non root node - we can't rely on
it pointing to the correct parent if we don't have the parent locked, but we can
use it to check if we have a parent that we need to have locked.

Also change btree_split() to set b->parent for the old root and sibling when
spliting the root.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>