bcache: Don't store prio bucket ptrs for journal in struct cache
If we write a journal entry when a device is (possibly temporarily) offline,
we're just going to write garbage for that slot - we want to preserve the prio
bucket ptr we wrote previously, possibly the one we found on journal replay.

Fix this by moving prio bucket ptrs the journal will use to c->journal; also
refactor some journalling code.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>